+++ forui/lib/forui.dart
@@ -4,7 +4,6 @@ library forui;
 export 'assets.dart';
 export 'foundation.dart';
 export 'localizations.dart';
-export 'properties.dart';
 export 'theme.dart';
 
 export 'widgets/accordion.dart';