+++ forui/lib/foundation.dart
@@ -3,8 +3,8 @@ library forui.foundation;
 
 export 'src/foundation/tappable.dart' show FTappable, FTappableData;
 export 'src/foundation/notifiers.dart';
-export 'src/properties/focusable_properties.dart';
-export 'src/properties/form_field_properties.dart';
+export 'src/foundation/focusable_properties.dart';
+export 'src/foundation/form_field_properties.dart';
 export 'src/foundation/focused_outline.dart';
 export 'src/foundation/portal/portal.dart';
 export 'src/foundation/portal/portal_shift.dart';