+++ forui/lib/properties.dart
@@ -1,5 +1,0 @@
-library forui.properties;
-
-export 'src/properties/focusable_properties.dart';
-export 'src/properties/form_field_properties.dart';
-export 'src/properties/popover_properties.dart';